fix: embed Elm form in svelte form so we do not need to bind reactive props
This fixes the issue in build mode when rerendering

Elm component while Virtual dom and dom

are not yet synced
